+++ backend/utils.js
@@ -7,7 +7,6 @@ const { COLLECTIONS } = require('../public/consts');
 
 const { CONFIG_KEYS, GEO_HASH_PRECISION } = require('./consts');
 const { wixData } = require('./elevated-modules');
-const { urlExists } = require('./members-data-methods');
 const elevatedGetSecretValue = auth.elevate(secrets.getSecretValue);
 
 /**
@@ -132,28 +131,6 @@ const normalizeUrlForComparison = url => {
   return url.toLowerCase().replace(/-\d+$/, '');
 };
 
-/**
- * Checks URL uniqueness for a member
- * @param {string} url - The URL to check
- * @param {string} memberId - The member ID to exclude from the check
- * @returns {Promise<Object>} Result object with isUnique boolean
- */
-async function checkUrlUniqueness(url, memberId) {
-  if (!url || !memberId) {
-    throw new Error('Missing required parameters: url and memberId are required');
-  }
-
-  try {
-    const trimmedUrl = url.trim();
-    const exists = await urlExists(trimmedUrl, memberId);
-
-    return { isUnique: !exists };
-  } catch (error) {
-    console.error('Error checking URL uniqueness:', error);
-    throw new Error(`Failed to check URL uniqueness: ${error.message}`);
-  }
-}
-
 async function getSecret(secretKey) {
   return await elevatedGetSecretValue(secretKey).value;
 }
@@ -197,7 +174,6 @@ module.exports = {
   isValidArray,
   normalizeUrlForComparison,
   queryAllItems,
-  checkUrlUniqueness,
   formatDateToMonthYear,
   isStudent,
   getAddressDisplayOptions,
